+++ settings.php
@@ -36,7 +36,7 @@ $settings = null;
 $ADMIN->add('themes', new admin_category('theme_adaptable', get_string('configtitle', 'theme_adaptable')));
 
 // Adaptable theme settings page.
-$asettings = new theme_adaptable_admin_settingspage_tabs(
+$asettings = new \theme_adaptable\admin_settingspage_tabs(
     'themesettingadaptable',
     get_string('configtabtitle', 'theme_adaptable'),
     402
@@ -76,6 +76,7 @@ if ($ADMIN->fulltree) {
     include(dirname(__FILE__) . '/settings/templates.php');
     include(dirname(__FILE__) . '/settings/user.php');
     include(dirname(__FILE__) . '/settings/custom_css.php');
+    include(dirname(__FILE__) . '/settings/custom_js.php');
 }
 $ADMIN->add('theme_adaptable', $asettings);
 require(dirname(__FILE__) . '/settings/importexport_settings.php');