+++ settings/colors.php
@@ -27,7 +27,7 @@ defined('MOODLE_INTERNAL') || die;
 
 // Colors section.
 if ($ADMIN->fulltree) {
-    $page = new theme_adaptable_admin_settingspage('theme_adaptable_color', get_string('colorsettings', 'theme_adaptable'));
+    $page = new \theme_adaptable\admin_settingspage('theme_adaptable_color', get_string('colorsettings', 'theme_adaptable'));
 
     $page->add(new admin_setting_heading(
         'theme_adaptable_color',