+++ settings/courses.php
@@ -25,7 +25,7 @@
 defined('MOODLE_INTERNAL') || die;
 
 if ($ADMIN->fulltree) {
-    $page = new theme_adaptable_admin_settingspage('theme_adaptable_course', get_string('coursesettings', 'theme_adaptable'));
+    $page = new \theme_adaptable\admin_settingspage('theme_adaptable_course', get_string('coursesettings', 'theme_adaptable'));
 
     $page->add(new admin_setting_heading(
         'theme_adaptable_course',