+++ settings/custom_css.php
@@ -28,14 +28,14 @@
 
 defined('MOODLE_INTERNAL') || die;
 
-// Custom CSS and JS section.
+// Custom CSS section.
 if ($ADMIN->fulltree) {
-    $page = new theme_adaptable_admin_settingspage('theme_adaptable_generic', get_string('customcssjssettings', 'theme_adaptable'));
+    $page = new \theme_adaptable\admin_settingspage('theme_adaptable_customcss', get_string('customcsssettings', 'theme_adaptable'));
 
     $page->add(new admin_setting_heading(
-        'theme_adaptable_generic',
-        get_string('genericsettingsheading', 'theme_adaptable'),
-        format_text(get_string('genericsettingsdescription', 'theme_adaptable'), FORMAT_MARKDOWN)
+        'theme_adaptable_customcss',
+        get_string('customcssjssettingsheading', 'theme_adaptable'),
+        format_text(get_string('customcsssettingsdescription', 'theme_adaptable'), FORMAT_MARKDOWN)
     ));
 
     // Custom CSS.
@@ -54,35 +54,6 @@ if ($ADMIN->fulltree) {
     $default = '';
     $setting = new admin_setting_configtextarea($name, $title, $description, $default);
     $setting->set_updatedcallback('theme_reset_all_caches');
-    $page->add($setting);
-
-    // Section for javascript to be added e.g. Google Analytics.
-    $name = 'theme_adaptable/jssection';
-    $title = get_string('jssection', 'theme_adaptable');
-    $description = get_string('jssectiondesc', 'theme_adaptable');
-    $default = '';
-    $setting = new admin_setting_configtextarea($name, $title, $description, $default);
-    $page->add($setting);
-
-    // Section for custom javascript, restricted by profile field.
-    $name = 'theme_adaptable/jssectionrestricted';
-    $title = get_string('jssectionrestricted', 'theme_adaptable');
-    $description = get_string('jssectionrestricteddesc', 'theme_adaptable');
-    $default = '';
-    $setting = new admin_setting_configtextarea($name, $title, $description, $default);
-    $page->add($setting);
-
-    $name = 'theme_adaptable/jssectionrestrictedprofilefield';
-    $title = get_string('jssectionrestrictedprofilefield', 'theme_adaptable');
-    $description = get_string('jssectionrestrictedprofilefielddesc', 'theme_adaptable');
-    $setting = new admin_setting_configtext($name, $title, $description, '', PARAM_RAW);
-    $page->add($setting);
-
-    $name = 'theme_adaptable/jssectionrestricteddashboardonly';
-    $title = get_string('jssectionrestricteddashboardonly', 'theme_adaptable');
-    $description = get_string('jssectionrestricteddashboardonlydesc', 'theme_adaptable');
-    $default = true;
-    $setting = new admin_setting_configcheckbox($name, $title, $description, $default, true, false);
     $page->add($setting);
 
     $asettings->add($page);