+++ settings/header_user.php
@@ -26,7 +26,7 @@
 defined('MOODLE_INTERNAL') || die;
 
 if ($ADMIN->fulltree) {
-    $page = new theme_adaptable_admin_settingspage('theme_adaptable_usernav', get_string('usernav', 'theme_adaptable'));
+    $page = new \theme_adaptable\admin_settingspage('theme_adaptable_usernav', get_string('usernav', 'theme_adaptable'));
 
     $page->add(new admin_setting_heading(
         'theme_adaptable_usernav',